put errors in their own file
